server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-pymongo==2.3.0
flask-jwt-extended==4.3.1
python-dotenv==0.19.0
pymongo==3.12.0
setuptools>=65.5.1
wheel>=0.38.0
numpy>=2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
transformers>=4.30.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.0.1
requests>=2.31.0